[response] deleted Publisher `validate` method
